Portfolio section and modal popup system
for my portfolio, i integrated a pop up event that displays a rectangle window that shows the details and image of the project/work. each project required their own data id which can be used to differentiate each popup in the javascript, using the get attribute. When the window is opened the modal style is on block while the click off cross is visible.